feat: Ajustado delete de tema, materia e prova

+++ server/src/service/MateriaService.ts
@@ -1,9 +1,14 @@
 import { Materia } from "../domain/entities/Materia";
 import { MateriaRepository } from "../domain/repositories/MateriaRepository";
+import { ProvaRepository } from "../domain/repositories/ProvaRepository";
+import { TemaRepository } from "../domain/repositories/TemaRepository";
 import { MateriaDTO } from "../dto/MateriaDTO";
+import { MateriaUtilizadaError } from "../errors/materia/MateriaUtilizadaError";
 
 export class MateriaService {
   private repository = new MateriaRepository();
+  private temaRepository = new TemaRepository();
+  private provaRepository = new ProvaRepository();
 
   async create(materiaDTO: MateriaDTO): Promise<Materia> {
 
@@ -33,6 +38,14 @@ export class MateriaService {
   async deleteById(id: number): Promise<void> {
 
     const materiaExistente = await this.repository.getById(id);
+
+    const temas = await this.temaRepository.getByMateriaId(id);
+
+    const provas = await this.provaRepository.getByMateriaId(id);
+
+    if (temas.length > 0 || provas.length > 0) {
+      throw new MateriaUtilizadaError();
+    }
 
     await this.repository.delete(materiaExistente);
   }
